refactor: server routes to /api endpoint
