+++ client/src/hooks/useBugs.tsx
@@ -5,7 +5,7 @@ import { Bug } from '../types/Bug';
 // Get all bugs for one project
 const useBugs = (projectId: number): QueryResult<Bug[]> => {
   return useQuery(projectId && ['projectbugs', projectId], () => {
-    return fetchRequest(`/bugs?projectId=${projectId}`);
+    return fetchRequest(`/api/bugs?projectId=${projectId}`);
   });
 };
