+++ client/src/hooks/useProject.tsx
@@ -5,7 +5,7 @@ import { Project } from '../types/Project';
 // Get details aboput a single project
 const useProject = (projectId: number): QueryResult<Project> => {
   return useQuery(projectId && ['project', projectId], () =>
-    fetchRequest(`/projects?projectId=${projectId}`)
+    fetchRequest(`/api/projects?projectId=${projectId}`)
   );
 };
