+++ client/src/hooks/useProjects.tsx
@@ -5,7 +5,7 @@ import { Project } from '../types/Project';
 // Get projects for a single user
 const useProjects = (userId: number): QueryResult<Project[]> => {
   return useQuery(userId && ['project', userId], () =>
-    fetchRequest(`/projects?userId=${userId}`)
+    fetchRequest(`/api/projects?userId=${userId}`)
   );
 };
