+++ client/src/hooks/useUser.tsx
@@ -5,7 +5,7 @@ import { User } from '../types/User';
 // Get details about one user
 const useUser = (userId: number): QueryResult<User> => {
   return useQuery(userId && ['user', userId], () =>
-    fetchRequest(`/user?userId=${userId}`)
+    fetchRequest(`/api/user?userId=${userId}`)
   );
 };
