+++ client/src/hooks/useUsers.tsx
@@ -5,7 +5,7 @@ import { User } from '../types/User';
 // Get all users for one project
 const useBugs = (projectId: number): QueryResult<User[]> => {
   return useQuery(projectId && ['projectusers', projectId], () => {
-    return fetchRequest(`/projectsUsers?projectId=${projectId}`);
+    return fetchRequest(`/api/projectsUsers?projectId=${projectId}`);
   });
 };
